Erro na função writerDatajson (entrada de dados no Dict; E escrita do arquivo)

+++ src/writer.py
@@ -2,42 +2,46 @@
 
 
 import os
+import sys
+import json
 
 
 def documentJson(filename):
-    caminho = '~/.dataset/'
-    # TODO: concatenar e criar o arquivo junto com a ext ".json"
-    arquivo = caminho + filename
+    caminho = '.dataset/'
+    arquivo = caminho + (filename + ".json")
     if not os.path.exists(caminho):
-        os.mkdir(caminho)
+        os.makedirs(caminho)
     if not os.path.exists(arquivo):
-        open(arquivo)
+        open(arquivo, "w+")
 
 
-def writerData(nameKeys, nameValues):
-    dictData = {}
+def documentCsv(filename):
+    caminho = '.dataset/'
+    arquivo = caminho + (filename + ".csv")
+    if not os.path.exists(caminho):
+        os.makedirs(caminho)
+    if not os.path.exists(arquivo):
+        open(arquivo, "w+")
+
+
+def writerDatajson(nameKeys, nameValues, filename):
+    # TODO: Função para criar a Dict que vai ser escrito no arquivo ".Json"
+    # dictData = {'nameKeys': 'nameValues'}
     arquivo = documentJson(filename)
-    keysValues = open(arquivo, "w+")
-    for keysValues in dictData:
-        dictData["nameKeys"] = nameValues
-        keysValues.write(dictData)
-        keysValues.close()
+    with open(arquivo, "w+") as abrir:
+            conteudo = abrir.readlines()
+            try:
+                for dictData in json.JSONEncoder().iterencode({"nameKeys": "nameValues"}):
+                    conteudo.append(str(dictData))
+                    abrir = open(arquivo, "w")
+                    abrir.writelines(conteudo)
+            except ValueError:
+               print('ERROR: Objeto não interou no Dict')
+            else:
+                abrir.close()
+                return arquivo
 
 
-# TODO: Escrever um looping para determinar a quantidade valores de  entra.
-
-
-def readData(dictData, nameKeys, nameValues):
-    if dictData == {nameKeys}:
-        readDataKeys = writerData(nameKeys)
-        readDataKeys.readlines()
-        return readDataKeys
-    elif dictData == {nameValues}:
-        readDataValues = writerData(nameValues)
-        readDataValues.readlines()
-        return readDataValues
-    else:
-        print("ERROR")
-        
-        # TODO: Escrever o modulo read || de forma imperativa escrever todas
-        # as funções no module writer.
+def writerDataCsv():
+    # TODO: Escrever função para gravar dados em arquivos csv
+    pass
